Update Partay.java
This class is not done at all. I copied over the stuff from Tester class and converted it to method form. This includes creating the array list, adding people to them, manually registering people through inputs while checking if there is over 100 members. It does not yet have auto-sort functionality for tables and seats but I added the method for searching a person's given information about tables and seats with an input. Next step is to add guests to table and give them a seat number.